Rewrite Sign and Text-Display Rendering Code to be more Flexible and Reusable (#2417)
* Revamp the sign layout system, remove several duplicate item.clone() calls.

* Fix issues where partial components just replaced each other.

* Rename partial -> inline for components.

* Push itemcomponent work so far.

* Conditional components, remove snapshot check for non snapshot rendering.

* Music disc song name display for signs and text-displays.

* Modify changelog and config.yml version.

+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/shop/display/virtual/packet/packetevents/PacketFactoryv1_20.java
@@ -152,7 +152,7 @@ public class PacketFactoryv1_20 implements PacketFactory<PacketWrapper<?>> {
     data.add(new EntityData<>(23, EntityDataTypes.ADV_COMPONENT, Util.getTextDisplay(shop, itemStack)));
     data.add(new EntityData<>(24, EntityDataTypes.INT, QuickShop.getInstance().getConfig().getInt("shop.text-display.line-width", 200)));
     data.add(new EntityData<>(25, EntityDataTypes.INT, QuickShop.getInstance().getConfig().getInt("shop.text-display.background-color", 1073741824)));
-    data.add(new EntityData<>(26, EntityDataTypes.INT, QuickShop.getInstance().getConfig().getInt("shop.text-display.text-opacity", -1)));
+    data.add(new EntityData<>(26, EntityDataTypes.BYTE, QuickShop.getInstance().getConfig().getByte("shop.text-display.text-opacity", (byte)-1)));
     data.add(new EntityData<>(27, EntityDataTypes.BYTE, Util.createTextDisplayFlags()));
 
     return new WrapperPlayServerEntityMetadata(id, data);
